datamind_final/project_final/backend/requirements.txt
# Core web framework
fastapi
uvicorn

# PDF and document processing
pypdf
PyPDF2
pdfplumber
python-docx
pytesseract
Pillow

# HTTP requests
requests

# Environment variable management
python-dotenv

# Supabase client
supabase